[Cleanup] Hide internal profiler APIs behind impl/profiler headers
Move metalium-internal profiler free functions and types out of the public
tt-metalium profiler headers so only ttnn/tt-train-facing symbols remain public.
