+++ tests/tt_metal/tt_metal/dispatch/dispatch_program/test_sub_device.cpp
@@ -23,7 +23,6 @@
 #include <tt-metalium/buffer.hpp>
 #include <tt-metalium/buffer_types.hpp>
 #include <tt-metalium/circular_buffer_config.hpp>
-#include <tt-metalium/tt_metal_profiler.hpp>
 #include "command_queue_fixture.hpp"
 #include <tt-metalium/kernel_types.hpp>
 #include "dispatch_test_utils.hpp"
