+++ tests/tt_metal/tt_metal/dispatch/dispatch_trace/test_sub_device.cpp
@@ -11,7 +11,6 @@
 #include <unordered_set>
 #include <vector>
 
-#include <tt-metalium/tt_metal_profiler.hpp>
 #include "command_queue_fixture.hpp"
 #include "dispatch_test_utils.hpp"
 #include "gtest/gtest.h"
