+++ tests/tt_metal/tt_metal/perf_microbenchmark/tensix/test_gathering.cpp
@@ -11,7 +11,6 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/program.hpp>
 #include <tt-metalium/tt_metal.hpp>
-#include <tt-metalium/tt_metal_profiler.hpp>
 #include <tt-metalium/distributed.hpp>
 
 namespace tt_metal = tt::tt_metal;
