+++ tests/ttnn/tracy/cpp/test_get_programs_perf_data.cpp
@@ -13,6 +13,7 @@
 #include <tt-metalium/tt_metal.hpp>
 #include "impl/context/metal_context.hpp"
 #include "impl/profiler/profiler_paths.hpp"
+#include "impl/profiler/tt_metal_profiler.hpp"
 #include <umd/device/types/cluster_descriptor_types.hpp>
 
 namespace tt::tt_metal::experimental {
