+++ tt_metal/api/tt-metalium/profiler_types.hpp
@@ -4,18 +4,9 @@
 
 #pragma once
 
-#include <cstdint>
-#include <tt-metalium/device_types.hpp>
-
 namespace tt::tt_metal {
 
 enum class ProfilerReadState { NORMAL, ONLY_DISPATCH_CORES, LAST_FD_READ };
 enum class ProfilerSyncState { INIT, CLOSE_DEVICE };
-enum class ProfilerDataBufferSource { L1, DRAM, DRAM_AND_L1 };
 
-struct DeviceProgramId {
-    uint32_t base_program_id = 0;
-    ChipId device_id = 0;
-    bool is_host_fallback_op = false;
-};
 }  // namespace tt::tt_metal
