+++ tt_metal/impl/profiler/tt_metal_profiler.hpp
@@ -1,0 +1,37 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstdint>
+#include <optional>
+#include <string>
+
+#include <tt-metalium/profiler_optional_metadata.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
+
+#include "profiler_types.hpp"
+
+namespace tt::tt_metal {
+class IDevice;
+
+namespace detail {
+
+void ClearProfilerControlBuffer(IDevice* device);
+
+void InitDeviceProfiler(IDevice* device);
+
+void ReadDeviceProfilerResults(
+    IDevice* device,
+    ProfilerReadState = ProfilerReadState::NORMAL,
+    const std::optional<ProfilerOptionalMetadata>& metadata = {});
+
+void SetDeviceProfilerDir(const std::string& output_dir = "");
+
+void FreshProfilerDeviceLog();
+
+DeviceProgramId DecodePerDeviceProgramID(uint32_t device_program_id);
+
+}  // namespace detail
+}  // namespace tt::tt_metal
