+++ tt_metal/programming_examples/profiler/test_dispatch_cores/test_dispatch_cores.cpp
@@ -10,7 +10,6 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/device.hpp>
-#include <tt-metalium/tt_metal_profiler.hpp>
 #include <tt-metalium/distributed.hpp>
 
 using namespace tt;
