+++ tt_metal/tools/profiler/tt_metal_tracy.hpp
@@ -4,7 +4,6 @@
 #pragma once
 
 #include "tracy_debug_zones.hpp"
-#include <tt_metal_profiler.hpp>
 #include "impl/context/metal_context.hpp"
 #include "distributed/mesh_device_impl.hpp"
 #include <tt_metal/impl/profiler/profiler_state.hpp>
